Database: dump() member should initialize the datafiles
The dump() member is constant, as it should be, but if called before one
of the other member that initialize the database, it will fail to return
any intervals.

This change makes the _files member mutable so that constant member
functions are able to initialize it.  These functions still maintain
logical constness as the database itself is not modified.

Signed-off-by: Shaun Ruffell <sruffell@sruffell.net>