Database: initialize datafiles in constructor
There were some paths through the database that could result in the
datafiles being improperly initialized. For example, it was possible to
call dump() on an initialized database that wouldn't return anything.

Since the vast majority of use cases involve loading all the Datafile, I
believe it is clearer to always initialize them at construction time.

Signed-off-by: Shaun Ruffell <sruffell@sruffell.net>